Add files generated by autogen.sh
This will make it possible to build the library without having to
install automake and autoconf.

It also means that autogen.sh will need to be rerun after each merge
with master.

Note that when doing, install.sh in the main repository must be
temporarily deleted or renamed, as otherwise autogen.sh would
auto-detect ../.. as the "aux dir" and generate some of the files there
rather than in this directory itself.