Add wx-specific copies of the generic files
These are needed in order to build PCRE using makefiles under MSW, so
add them under wx-specific directory that will be specified as an extra
include path only when building PCRE as part of wx.

Currently these are just unmodified copies of the original files, but
they will be changed in the upcoming commits.